adds jest to types, moves scripts to src
