+++ .prettierignore
@@ -1,7 +1,7 @@
 package-lock.json
 .next
 node_modules/
-scripts/
+src/scripts/
 .github/workflows/
 templates/css/bootstrap.min.css
 templates/js/paged.polyfill.min.js
