+++ src/__tests__/api/pdf/experiment-safety-pdf.spec.ts
@@ -3,9 +3,9 @@ import { setTimeout } from 'timers/promises';
 
 import request from 'supertest';
 
-import extractPDFText from '../../../../scripts/text-extraction';
 import app from '../../../app';
 import { getTotalPages } from '../../../pdf';
+import extractPDFText from '../../../scripts/text-extraction';
 import { generateTmpPath } from '../../../util/fileSystem';
 import testPayloads from '../../fixtures/pdf-payloads.json';
