+++ src/scripts/text-extraction/index.ts
