+++ src/scripts/text-extraction/lib/collection-state.js
