+++ src/scripts/text-extraction/lib/encoding/adobe-glyph-list.js
