+++ src/scripts/text-extraction/lib/encoding/mac-expert-encoding.js
