+++ src/scripts/text-extraction/lib/encoding/mac-roman-encoding.js
