+++ src/scripts/text-extraction/lib/encoding/standard-encoding.js
