+++ src/scripts/text-extraction/lib/encoding/symbol-encoding.js
