+++ src/scripts/text-extraction/lib/encoding/win-ansi-encoding.js
