+++ src/scripts/text-extraction/lib/font-decoding.js
