+++ src/scripts/text-extraction/lib/multi-dict-helper.js
