+++ src/scripts/text-extraction/lib/pdf-interpreter.js
