+++ src/scripts/text-extraction/lib/placements-extraction.js
