+++ src/scripts/text-extraction/lib/standard-fonts-dimensions.js
