+++ src/scripts/text-extraction/lib/text-extraction.js
