+++ src/scripts/text-extraction/lib/transformations.js
